+++ lib/src/players/launch_player_function.dart
@@ -4,4 +4,4 @@ import '../config/config.dart';
 import '../player_arguments.dart';
 
 
-typedef LaunchPlayerFunction = FutureOr<void> Function(Config config, PlayerArguments playerArguments);
+typedef LaunchPlayerFunction = Future<void>? Function(Config config, PlayerArguments playerArguments);