+++ userscripts/anime365.meta.js
@@ -1,16 +1,18 @@
 // ==UserScript==
 // @name         Anime365 play via PotPlayer
 // @namespace    https://znnme.eu.org/
-// @version      3.0.0
+// @version      3.1.0
 // @description  Add play via PotPlayer button to Anime365 website.
 // @downloadURL  https://raw.githubusercontent.com/Zekfad/player_launcher/master/userscripts/anime365.user.js
 // @updateURL    https://raw.githubusercontent.com/Zekfad/player_launcher/master/userscripts/anime365.meta.js
 // @author       Zekfad
-// @match        https://smotret-anime.com/*
-// @match        https://smotret-anime.online/*
 // @match        https://anime365.ru/*
 // @match        https://hentai365.ru/*
+// @match        https://smotret-anime.com/*
+// @match        https://smotret-anime.online/*
+// @match        https://smotret-anime.org/*
 // @icon         https://www.google.com/s2/favicons?sz=64&domain=anime365.ru
+// @icon64       https://www.google.com/s2/favicons?sz=64&domain=anime365.ru
 // @grant        none
 // @run-at document-end
 // ==/UserScript==